+++ package.json
@@ -1,8 +1,8 @@
 {
-  "name": "appframe",
-  "version": "0.11.2",
-  "description": "Basic framework app object for dependency injection, event propagation, and error handling.",
-  "main": "appframe.js",
+  "name": "spawnpoint",
+  "version": "1.0.0",
+  "description": "Agnostic JS framework that empowers devs to focus on quickly building apps, rather than focusing on application config, health-checks, application structure, or architecture to build a 12 factor app in Docker.",
+  "main": "spawnpoint.js",
   "scripts": {
     "test": "npm run lint",
     "lint": "eslint .",
@@ -10,7 +10,7 @@
   },
   "repository": {
     "type": "git",
-    "url": "https://github.com/nodecraft/appframe.js"
+    "url": "https://github.com/nodecraft/spawnpoint"
   },
   "keywords": [
     "app",
@@ -18,12 +18,12 @@
     "events",
     "errors"
   ],
-  "author": "Nodecraft Inc.",
+  "author": "Nodecraft, Inc.",
   "license": "MIT",
   "bugs": {
-    "url": "https://github.com/nodecraft/appframe.js/issues"
+    "url": "https://github.com/nodecraft/spawnpoint/issues"
   },
-  "homepage": "https://github.com/nodecraft/appframe.js",
+  "homepage": "https://github.com/nodecraft/spawnpoint",
   "dependencies": {
     "async": "*",
     "chalk": "^2.4.1",
